chunk_renderer refactor

+++ src/renderer/gl/texture.cpp
@@ -40,9 +40,9 @@ TextureAtlas::TextureAtlas(const std::string_view filepath, const glm::uvec2& su
 }
 
 void TextureAtlas::getSubTextureUVs(
-    const glm::ivec2& position, glm::vec2* outMinUV, glm::vec2* outMaxUV)
+    const glm::uvec2& position, glm::vec2* outMinUV, glm::vec2* outMaxUV)
 {
     glm::vec2 toUVCoord = static_cast<glm::vec2>(m_subTextureSize) / static_cast<glm::vec2>(m_size);
     *outMinUV = static_cast<glm::vec2>(position) * toUVCoord;
-    *outMaxUV = static_cast<glm::vec2>(position + 1) * toUVCoord;
+    *outMaxUV = static_cast<glm::vec2>(position + static_cast<uint32_t>(1)) * toUVCoord;
 }
